+++ app/Http/Controllers/Controller.php
@@ -69,10 +69,12 @@ class Controller extends BaseController
             'verbs' => [],
             'http_status_code' => $http_status_code,
             'headers' => [
-                'Access-Control-Allow-Origin' => '*',
-                'Access-Control-Allow-Header' => 'X-Requested-With, Origin, Content-Type, Accept, Authorization',
-                'Access-Control-Allow-Methods' => implode(', ', array_keys($verbs)) . ', OPTIONS',
-                'Content-Type' => 'application/json'
+                'Content-Security-Policy' => 'default-src \'none\'',
+                'Strict-Transport-Security' => 'max-age=31536000;',
+                'Content-Type' => 'application/json',
+                'Content-Language' => app()->getLocale(),
+                'Referrer-Policy' => 'origin-when-cross-origin',
+                'X-Content-Type-Options' => 'nosniff'
             ]
         ];