+++ app/Utilities/Header.php
@@ -23,7 +23,14 @@ class Header
 
     public function __construct()
     {
-        $this->headers = [];
+        $this->headers = [
+            'Content-Security-Policy' => 'default-src \'none\'',
+            'Strict-Transport-Security' => 'max-age=31536000;',
+            'Content-Type' => 'application/json',
+            'Content-Language' => app()->getLocale(),
+            'Referrer-Policy' => 'origin-when-cross-origin',
+            'X-Content-Type-Options' => 'nosniff'
+        ];
     }
 
     /**
@@ -42,14 +49,17 @@ class Header
         int $total_count
     ): Header
     {
-        $this->headers = [
-            'X-Count' => $count,
-            'X-Total-Count' => $total_count,
-            'X-Offset' => $pagination['offset'],
-            'X-Limit' => $pagination['limit'],
-            'X-Link-Previous' => $pagination['links']['previous'],
-            'X-Link-Next' => $pagination['links']['next']
-        ];
+        $this->headers = array_merge(
+            $this->headers,
+            [
+                'X-Count' => $count,
+                'X-Total-Count' => $total_count,
+                'X-Offset' => $pagination['offset'],
+                'X-Limit' => $pagination['limit'],
+                'X-Link-Previous' => $pagination['links']['previous'],
+                'X-Link-Next' => $pagination['links']['next']
+            ]
+        );
 
         return $this;
     }
@@ -61,8 +71,13 @@ class Header
      */
     public function item(): Header
     {
-        $this->add('X-Total-Count', 1);
-        $this->add('X-Count', 1);
+        $this->headers = array_merge(
+            $this->headers,
+            [
+                'X-Total-Count' => 1,
+                'X-Count' => 1
+            ]
+        );
 
         return $this;
     }